fix: audit round — Flutter correctness, service hygiene, docs accuracy
Flutter:
- restore form validation (an if(true) short-circuited it; unvalidated data
  was persisted and the required-fields dialog was unreachable)
- rewire generated l10n (AppLocalizations delegates + typed translation());
  labels no longer render as empty strings — every key used is present in
  the generated localizations
- await Persistence.initPersistence() (theme/user no longer race the load)
- password fields start obscured; remove dead isMyButtonEnabled fields
- unwrap nested MaterialApps (welcome/measurement/guide) — theme + navigator
  stack now behave; remove stray second main() in guide_screen
- replace cleartext third-party Google-logo hotlink (pngimg.com) with a
  Material icon placeholder (brand asset tracked for PRD)
- interpolate persisted user data in verification copy (removes hardcoded
  personal email/phone); dedupe 3 copy-pasted Countdown widgets into
  src/shared/countdown.dart; remove unused OTP controllers + debug print
- android: add INTERNET permission to the main manifest (release builds
  could make zero network calls); applicationId io.cuesoft.apparule
- pubspec: real description; drop 5 unused deps (http, animations,
  flutter_animate, font_awesome_flutter, cupertino_icons)
- remove dead files: empty profile_screen.dart, unimported my_app_bar.dart,
  no-op widget_test.dart

api/measure: remove 4.9MB of unreferenced test images from git; add
.dockerignore (tests/images no longer ship in the image); PORT-aware
healthcheck; pin requirements to the resolved image versions + add
requirements-dev.txt (pytest)

api/common: CORS now emits Vary: Origin alongside the reflected allow-origin

web: remove template svgs; real README (scaffold status made explicit)

docs/README: Go 1.26 / Python 3.12 to match go.mod and images; root-.env
claim fixed; helm/docker rows marked planned

Verified: go build + go test green; dart sources balance-checked and l10n
keys cross-checked (no flutter toolchain locally — analyze deferred).
