+++ api/measure/.dockerignore
@@ -1,0 +1,7 @@
+__pycache__/
+*.pyc
+.pytest_cache/
+tests/
+.env
+.env.*
+*.md
