+++ api/measure/Dockerfile
@@ -16,5 +16,5 @@ USER appuser
 ENV PORT=8081
 EXPOSE 8081
 HEALTHCHECK --interval=30s --timeout=5s --start-period=40s --retries=3 \
-  CMD python -c "import urllib.request,sys; sys.exit(0 if urllib.request.urlopen('http://127.0.0.1:8081/health').status==200 else 1)"
+  CMD python -c "import os,urllib.request,sys; sys.exit(0 if urllib.request.urlopen('http://127.0.0.1:'+os.environ.get('PORT','8081')+'/health').status==200 else 1)"
 CMD ["sh","-c","uvicorn app.main:app --host 0.0.0.0 --port ${PORT:-8081}"]
