+++ api/measure/requirements.txt
@@ -1,6 +1,6 @@
-fastapi>=0.115,<0.116
-uvicorn[standard]>=0.30,<0.35
-python-multipart>=0.0.9
-mediapipe>=0.10.14,<0.11
-opencv-python-headless>=4.10,<4.12
-numpy>=1.26,<2
+fastapi==0.115.14
+uvicorn[standard]==0.34.3
+python-multipart==0.0.32
+mediapipe==0.10.18
+opencv-python-headless==4.11.0.86
+numpy==1.26.4
