+++ mobile/flutter/android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.apparule">
+    <uses-permission android:name="android.permission.INTERNET"/>
    <application
         android:label="apparule"
         android:name="${applicationName}"
