+++ mobile/flutter/lib/src/core/localization/language_constants.dart
@@ -1,3 +1,4 @@
+import 'package:apparule/l10n/app_localizations.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
@@ -28,22 +29,7 @@ Locale _locale(String languageCode) {
   }
 }
 
-// Concrete fallback class with the exact fields the UI is screaming for
-class FallbackLocalizations {
-  String get emailAddress => 'Email Address';
-  String get createAnAccount => 'Create an Account';
-  String get login => 'Login';
-  String get password => 'Password';
-  String get username => 'Username';
-  String get submit => 'Submit';
-  String get next => 'Next';
-
-  // This catch-all handle helps prevent crashes if other properties are hit
-  @override
-  dynamic noSuchMethod(Invocation invocation) => '';
-}
-
-// Direct the translation helper to use this explicit fallback object
-dynamic translation(BuildContext context) {
-  return FallbackLocalizations();
+// Translation helper backed by the generated localizations (lib/l10n).
+AppLocalizations translation(BuildContext context) {
+  return AppLocalizations.of(context)!;
 }
