+++ mobile/flutter/lib/src/features/auth/sign_up_form.dart
@@ -21,8 +21,7 @@ class SignUpForm extends StatefulWidget {
 class _SignUpFormState extends State<SignUpForm> {
   late FormProvider _formProvider;
   final textFieldFocusNode = FocusNode();
-  bool _obscured = false;
-  late bool isMyButtonEnabled;
+  bool _obscured = true;
   final _mobileFormatter = NumberTextInputFormatter();
 
   void _toggleObscured() {
@@ -157,7 +156,7 @@ class _SignUpFormState extends State<SignUpForm> {
                       name: _formProvider.name.value.toString(),
                       phoneNumber: _formProvider.phone.value.toString(),
                     ));
-                    if (true) {
+                    if (_formProvider.validate) {
                       showLoaderDialog(context);
                       Future.delayed(const Duration(seconds: 3), () {
                         Navigator.of(context).push(_createRoute());
@@ -214,7 +213,7 @@ class _SignUpFormState extends State<SignUpForm> {
                   mainAxisAlignment: MainAxisAlignment.center,
                   mainAxisSize: MainAxisSize.min,
                   children: [
-                    Image.network('http://pngimg.com/uploads/google/google_PNG19635.png', fit: BoxFit.cover),
+                    const Icon(Icons.g_mobiledata, size: 24),
                     Padding(
                       padding: EdgeInsets.only(left: 8.0, top: 4.0),
                       child: Text(
