+++ mobile/flutter/lib/src/features/auth/verify_email.dart
@@ -1,5 +1,6 @@
 import 'package:apparule/src/app/home_screen.dart';
 import 'package:apparule/src/features/auth/reset_password.dart';
+import 'package:apparule/src/shared/countdown.dart';
 import 'package:flutter/material.dart';
 import 'package:sms_autofill/sms_autofill.dart';
 
@@ -13,8 +14,6 @@ class VerifyEmailForPasswordReset extends StatefulWidget {
 class _VerifyEmailForPasswordResetState extends State<VerifyEmailForPasswordReset> with SingleTickerProviderStateMixin {
   AnimationController? _animationController;
   int levelClock = 2 * 60;
-  final TextEditingController _emailController = TextEditingController();
-  final TextEditingController _otpController = TextEditingController();
 
   @override
   void initState() {
@@ -60,7 +59,7 @@ class _VerifyEmailForPasswordResetState extends State<VerifyEmailForPasswordRese
                     style: TextStyle(fontWeight: FontWeight.bold, fontSize: 22, color: Theme.of(context).colorScheme.onBackground),
                   ),
                   Text(
-                    "Check your Email inbox for the code sent to baasit.quadri@cuesoft.io. Enter the code below to complete the verification",
+                    "Check your Email inbox for the code sent to ${Persistence.getUser()?.email ?? 'your email'}. Enter the code below to complete the verification",
                     style: TextStyle(fontSize: 16, color: Theme.of(context).colorScheme.onBackground),
                   ),
                 ],
@@ -142,20 +141,3 @@ class _VerifyEmailForPasswordResetState extends State<VerifyEmailForPasswordRese
   }
 }
 
-class Countdown extends AnimatedWidget {
-  Countdown({Key? key, required this.animation}) : super(key: key, listenable: animation);
- final Animation<int> animation;
-
-  @override
-  build(BuildContext context) {
-    Duration clockTimer = Duration(seconds: animation.value);
-    String timerText = '${clockTimer.inMinutes.remainder(60).toString()}:${clockTimer.inSeconds.remainder(60).toString().padLeft(2, '0')}';
-    return Text(
-      timerText,
-      style: TextStyle(
-        fontSize: 18,
-        color: Theme.of(context).colorScheme.onBackground,
-      ),
-    );
-  }
-}
