+++ mobile/flutter/lib/src/shared/countdown.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+/// Shared mm:ss countdown text driven by an [Animation] of remaining seconds.
+class Countdown extends AnimatedWidget {
+  const Countdown({Key? key, required this.animation})
+      : super(key: key, listenable: animation);
+
+  final Animation<int> animation;
+
+  @override
+  Widget build(BuildContext context) {
+    Duration clockTimer = Duration(seconds: animation.value);
+    String timerText =
+        '${clockTimer.inMinutes.remainder(60).toString()}:${clockTimer.inSeconds.remainder(60).toString().padLeft(2, '0')}';
+    return Text(
+      timerText,
+      style: TextStyle(
+        fontSize: 18,
+        color: Theme.of(context).colorScheme.onBackground,
+      ),
+    );
+  }
+}
